Fix AndroidProxy compilation for non-Android
